Refactoring and adding new interfaces and structures
* Addition of the `using` directives for ILGPU and System.
* Update of `Dimension.cs` to include new references and use `InternalDimension`.
* Modification of `AcceleratorExecutor.cs` to use `TCoordinates`.
* Addition of the `ShapeIndex` attribute in `OperationInfo.cs`.
* Use of `Shape.Rank` for allocations in `Result.cs`.
* Addition of a default `ShapeIndex` in `ScriptBuilder.cs`.
* Replacement of `StaticArray` by `IStaticArray` and related structures.
* Simplification of `IShape` by inheriting from `IShapeBase`.
* Change the enumerations `SharedOpCode` and `SharedReduceCode` from `short` to `byte`, adding `SharedBroadcastCode`.
* Addition of static broadcast methods in `Shape.cs` and renaming of properties.
* Length checks using `Rank` in `ShapeIndex.cs`.
* Addition of the files `IShapeBase.cs`, `InternalDimension.cs`, and `InternalIndices.cs`.
* Introduction of the interface `IInternalShape<TXD>` and related structures in `InternalShape.cs`.
* Creation of `InternalShapeIndices<TShape, TIndices, TXD>` in `InternalShapeIndices.cs`.
* Addition of the interface `IInternalTensor<TShape, TIndices, TXD>` and the associated structure in `InternalTensor.cs`.
* Introduction of the interface `IShapeIndices` in `IShapeIndices.cs`.
* Addition of the interface `IXD` and structures `_1D`, `_2D`, `_3D` in `_XD.cs`.

+++ SharpGrad/Operators/SharedOpCode.cs
@@ -6,7 +6,7 @@ namespace SharpGrad.Operators
     /// Shared operators code.
     /// </summary>
     /// <remarks>It is a bit field short[0:2].</remarks>
-    public enum SharedOpCode : short // short[0:2]
+    public enum SharedOpCode : byte // short[0:2]
     {
         /// <summary>
         /// IsCommutative is set for commutative <see cref="SharedOpCode"/>
